Refactor ExamplePage component; clean up imports and formatting for better readability
